Rebuilt Project
Rebuilt the project with self-define maths and geometry classes.

I still use SFML for the actual drawing but have now implemented a basic mesh object (needs to be replaced to accodomate a vertex buffer + index buffer instead of a list of 'triangle' objects, but for now will suffice).

The goal of the rewrite was to make hte libraries more meticilous and math oriented so I may implement quaternions relatively soon.